Move sharing status indicator out of the call crate and into collab_ui in order so that the model doesn't depend on the view
