+++ crates/collab_ui/src/sharing_status_indicator.rs
@@ -1,10 +1,30 @@
+use call::ActiveCall;
 use gpui::{
     color::Color,
     elements::{MouseEventHandler, Svg},
-    Appearance, Element, ElementBox, Entity, MouseButton, RenderContext, View,
+    Appearance, Element, ElementBox, Entity, MouseButton, MutableAppContext, RenderContext, View,
 };
+use settings::Settings;
 
-use crate::ToggleScreenSharing;
+use crate::collab_titlebar_item::ToggleScreenSharing;
+
+pub fn init(cx: &mut MutableAppContext) {
+    let active_call = ActiveCall::global(cx);
+
+    let mut status_indicator = None;
+    cx.observe(&active_call, move |call, cx| {
+        if let Some(room) = call.read(cx).room() {
+            if room.read(cx).is_screen_sharing() {
+                if status_indicator.is_none() && cx.global::<Settings>().show_call_status_icon {
+                    status_indicator = Some(cx.add_status_bar_item(|_| SharingStatusIndicator));
+                }
+            } else if let Some((window_id, _)) = status_indicator.take() {
+                cx.remove_status_bar_item(window_id);
+            }
+        }
+    })
+    .detach();
+}
 
 pub struct SharingStatusIndicator;
